+++ BNS.Application/Features/Account/LoginNoPaswordCommand.cs
@@ -20,7 +20,7 @@ using BNS.Utilities;
 using BNS.Domain.Responses;
 
 namespace BNS.Service.Features
-{ 
+{
     public class LoginNoPaswordCommand : IRequestHandler<LoginNoPasswordRequest, ApiResult<LoginResponse>>
     {
         protected readonly IStringLocalizer<SharedResource> _sharedLocalizer;
@@ -47,14 +47,14 @@ namespace BNS.Service.Features
                 return response;
             }
             var userCompanys = await _unitOfWork.JM_AccountCompanyRepository.GetAsync(s => s.UserId == user.Id);
-            if (!userCompanys.Any(s => s.Status == (int)EUserStatus.ACTIVE))
+            if (!userCompanys.Any(s => s.Status == EUserStatus.ACTIVE))
             {
                 response.errorCode = EErrorCode.Failed.ToString();
                 response.title = _sharedLocalizer[LocalizedBackendMessages.User.MSG_UserOrPasswordNotCorrect];
                 return response;
             }
 
-            var userCompany = await userCompanys.Where(s => s.IsDefault && s.Status == (int)EUserStatus.ACTIVE).FirstOrDefaultAsync();
+            var userCompany = await userCompanys.Where(s => s.IsDefault && s.Status == EUserStatus.ACTIVE).FirstOrDefaultAsync();
 
             var roles = new List<string>();
             //if (userCompany.IsMainAccount)
